Renamed factory package to factorymethod
