+++ src/com/pattern/factorymethod/ChicagoPizzaStore.java
@@ -1,4 +1,4 @@
-package com.pattern.factory;
+package com.pattern.factorymethod;
 
 public class ChicagoPizzaStore extends PizzaStore {
