+++ src/com/pattern/factorymethod/ChicagoStyleCheesePizza.java
@@ -1,4 +1,4 @@
-package com.pattern.factory;
+package com.pattern.factorymethod;
 
 public class ChicagoStyleCheesePizza extends Pizza{
