+++ src/com/pattern/factorymethod/ChicagoStyleVeggiePizza.java
@@ -1,16 +1,18 @@
-package com.pattern.factory;
+package com.pattern.factorymethod;
 
-public class ChicagoStyleClamPizza extends Pizza {
+public class ChicagoStyleVeggiePizza extends Pizza {
 	
-	
-	public ChicagoStyleClamPizza() {
+	public ChicagoStyleVeggiePizza() {
 		
 		name = "Chicago Style Deep Dish Cheese Pizza";
 		dough = "Extra Thick Crust Dough";
 		sauce = "Plum Tomato Sauce";
 		
 		toppings.add("Shredded Mozzarella Cheese");
-		toppings.add("Clams");
+		toppings.add("Bell Peppers");
+		toppings.add("Mushrooms");
+		toppings.add("Onions");
+		toppings.add("Black Olives");
 	}
 	
 	void cut() {
