+++ src/com/pattern/factorymethod/NYPizzaStore.java
@@ -1,4 +1,4 @@
-package com.pattern.factory;
+package com.pattern.factorymethod;
 
 public class NYPizzaStore extends PizzaStore {
