+++ src/com/pattern/factorymethod/NYStyleClamPizza.java
@@ -1,4 +1,4 @@
-package com.pattern.factory;
+package com.pattern.factorymethod;
 
 public class NYStyleClamPizza extends Pizza {
