+++ src/com/pattern/factorymethod/NYStylePepperoniPizza.java
@@ -1,4 +1,4 @@
-package com.pattern.factory;
+package com.pattern.factorymethod;
 
 public class NYStylePepperoniPizza extends Pizza {
